requirements.txt
# DBT and BigQuery dependencies
dbt-bigquery==1.8.2
dbt-core==1.8.4

# Google Cloud dependencies
google-cloud-bigquery==3.25.0
google-cloud-secret-manager==2.20.2
google-cloud-monitoring==2.21.0

# Data processing
pandas==2.2.2
numpy==2.1.1

# Web framework for Cloud Run endpoints
flask==3.0.3
gunicorn==23.0.0

# Monitoring and logging
structlog==24.4.0
python-json-logger==2.0.7

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.2
requests==2.32.3